WIF-48: extract writeback.go and status.go from delegatedauth.go
Pure file move, no behavior change, no API change.

delegatedauth.go was 969 lines doing 5 jobs: component struct/init,
AddInstance, background refresh, config writeback, and status reporting.
Extracted the two largest self-contained concerns:

- writeback.go (~250 lines): writeAPIKeyToTarget,
  mergeIntoAdditionalEndpoints, mergeIntoAdditionalEndpointsList,
  resolveTargetSite, fallbackTargetInstance, updateConfigWithAPIKey.
  This is the flat-key path config writeback code. It only serves
  callers with SkipConfigWriteback=false.

- status.go (~125 lines): Name, Section, JSON, Text, HTML,
  populateStatusInfo, and the templatesFS embed.

delegatedauth.go is now ~620 lines, focused on the component struct,
AddInstance, initialization, background refresh, and the provider
registry. The writeback code is now skippable in review by anyone who
only cares about the directive (Provider) path.
